Rename GraphBlockPacker to BlockPacker
